+++ priv/static/phoenix_live_view.esm.js
@@ -1975,7 +1975,7 @@ var JS = {
   },
   exec_push(eventType, phxEvent, view, sourceEl, el, args) {
     let { event, data, target, page_loading, loading, value } = args;
-    let pushOpts = { page_loading: !!page_loading, loading, value };
+    let pushOpts = { loading, value, target, page_loading: !!page_loading };
     let targetSrc = eventType === "change" ? sourceEl.form : sourceEl;
     let phxTarget = target || targetSrc.getAttribute(view.binding("target")) || targetSrc;
     view.withinTargets(phxTarget, (targetView, targetCtx) => {
@@ -2791,14 +2791,14 @@ var View = class {
     let cid = el.getAttribute && el.getAttribute(PHX_COMPONENT);
     return cid ? parseInt(cid) : null;
   }
-  targetComponentID(target, targetCtx) {
+  targetComponentID(target, targetCtx, opts = {}) {
     if (isCid(targetCtx)) {
       return targetCtx;
     }
     let cidOrSelector = target.getAttribute(this.binding("target"));
     if (isCid(cidOrSelector)) {
       return parseInt(cidOrSelector);
-    } else if (targetCtx && cidOrSelector !== null) {
+    } else if (targetCtx && (cidOrSelector !== null || opts.target)) {
       return this.closestComponentID(targetCtx);
     } else {
       return null;
@@ -2862,7 +2862,7 @@ var View = class {
       type,
       event: phxEvent,
       value: this.extractMeta(el, meta, opts.value),
-      cid: this.targetComponentID(el, targetCtx)
+      cid: this.targetComponentID(el, targetCtx, opts)
     });
   }
   pushFileProgress(fileEl, entryRef, progress, onReply = function() {