+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Fix diff on response failing to await for active transitions in certain cases
   - Fix `phx-click-away` not respecting `phx-target`
   - Fix "disconnect" broadcast failing to failsafe refresh the page
+  - Fix `JS.push` with `:target` failing to send to correct component in certain cases
 
 ### Deprecations
    - Deprecate `Phoenix.LiveView.get_connect_info/1` in favor of `get_connect_info/2`